refactor: divide director files
